Fix thumbnails in Python 3
b64encode() returns a string in Python 2, but in Python 3, it returns
bytes. This needs to be converted into a string before formatting, or
you end up with b'' in the HTML, which breaks the thumbnails.

Presumably this breaks the thumbnails in Python 2. I haven't yet looked
into the best way to get it to work in both versions.